Move a bunch of tests to a separate directory specifically for the corenlp client
